# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OCMS — Construction ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UploadUrlRequest' }
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadUrlResponse' }
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema: { type: string, format: binary }
        '404':
          description: Object not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '500':
          description: Failed to serve object.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  /projects/{projectId}/issues:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      tags: [Issues]
      operationId: listProjectIssues
      responses:
        '200':
          description: Issues
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Issue' }
    post:
      tags: [Issues]
      operationId: createIssue
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IssueInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Issue' }
  /issues/{issueId}:
    parameters:
      - name: issueId
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Issues]
      operationId: updateIssue
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IssueUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Issue' }
...
  /dsr-rates:
    get:
      tags: [DSR]
      operationId: listDsrRates
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: trade
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: cityTier
          in: query
          schema: { type: string }
      responses:
        '200':
          description: DSR rates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DsrRate' }
    post:
      tags: [DSR]
      operationId: createDsrRate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DsrRateInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DsrRate' }
  /dsr-rates/{rateId}:
    parameters:
      - name: rateId
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [DSR]
      operationId: updateDsrRate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DsrRateUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DsrRate' }
    delete:
      tags: [DSR]
      operationId: deleteDsrRate
      responses:
        '204':
          description: Deleted
...
components:
  parameters:
...
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata: { $ref: '#/components/schemas/UploadUrlRequest' }
...
    IssueSeverity:
      type: string
      enum: [low, medium, high, critical]
...
    Issue:
      type: object
      required: [id, projectId, dprId, title, description, severity, status, raisedById, assignedToId, raisedAt, resolvedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        dprId: { type: [string, 'null'] }
        title: { type: string }
        description: { type: [string, 'null'] }
        severity: { $ref: '#/components/schemas/IssueSeverity' }
        status: { type: string }
        raisedById: { type: [string, 'null'] }
        assignedToId: { type: [string, 'null'] }
        raisedAt: { type: string, format: date-time }
        resolvedAt: { type: [string, 'null'], format: date-time }
    IssueInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        severity: { $ref: '#/components/schemas/IssueSeverity' }
        dprId: { type: string }
        assignedToId: { type: string }
    IssueUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        severity: { $ref: '#/components/schemas/IssueSeverity' }
        status: { type: string }
        assignedToId: { type: string }
...
    DsrRate:
      type: object
      required: [id, code, description, trade, unit, state, cityTier, rate, effectiveYear, source, createdById, createdAt, updatedAt]
      properties:
        id: { type: string }
        code: { type: string }
        description: { type: string }
        trade: { type: string }
        unit: { type: string }
        state: { type: string }
        cityTier: { type: string }
        rate: { type: number }
        effectiveYear: { type: integer }
        source: { type: string }
        createdById: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DsrRateInput:
      type: object
      required: [code, description, trade, unit, state, rate]
      properties:
        code: { type: string }
        description: { type: string }
        trade: { type: string }
        unit: { type: string }
        state: { type: string }
        cityTier: { type: string }
        rate: { type: number }
        effectiveYear: { type: integer }
        source: { type: string }
    DsrRateUpdate:
      type: object
      properties:
        description: { type: string }
        trade: { type: string }
        unit: { type: string }
        state: { type: string }
        cityTier: { type: string }
        rate: { type: number }
        effectiveYear: { type: integer }
        source: { type: string }
...